Add custom oAuth view to project
I've added some components, all of them responsible to oAuth and normal
authentication process. They used the accounts package, but they are a
substitution to the previous ugly interface.

The auth component compose of the buttons at the navigation bar, that
will redirect to login, or register view. Also, when user is logged, it
shows option to logout, as well the name of the user.

The login component compose the login view. When an user already
registered load its informations on the login/password options. It have
link to all other sections.

The register component compose the register view. When an user not
registered load its informations to sign up for the site. Also have a
link to other sections.

The password component compose the 'miss your password' view. When an
user forget its password and ask for an e-mail to re-register its
password. Also have link to others sections.

Note that the login and register have the oAuth social login button.
All of them does the same thing, log the user. The accounts system,
automatically handle the case when user isn't registered.