Add inteligent configuration of GoogleMaps API
I've added a function to load on every component that uses the maps
directive. But it didn't work to load the API keys. With that, I'm
waiting for this provider to improve. While I'm waiting, I'm gonna use
the unsafe method of providing the API key on a script call.